+++ src/main/java/edu/mioib/qaplocalsearch/model/StateEvaluation.java
@@ -4,6 +4,6 @@ import lombok.Value;
 
 @Value
 public class StateEvaluation {
-	int evaluation;
+	long evaluation;
 	int[] state;
 }
